requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
fastai==2.7.14
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.13.4
Flask==3.0.2
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.12
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchain-weaviate==0.0.1rc5
langcodes==3.3.0
langsmith==0.1.29
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pipfreeze==2.0.3
preshed==3.0.9
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
simsimd==3.9.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
srsly==2.4.8
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
wasabi==1.1.2
weasel==0.3.4
weaviate-client==4.5.7
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
langchain
ollama==0.3.3
pdfplumber==0.11.4